Add ManageCoursePage to router.
In this commit the ManageCourse page added to the router.
The course with slug url is before the course one
because if the switch find a match it will return and if the
course one was above the course with slug then
the slug one would have never been visited.